Use int() over ctypes.u_long since it will do the job we want

+++ src/middlewared/middlewared/plugins/crypto.py
@@ -1,4 +1,3 @@
-import ctypes
 import dateutil
 import dateutil.parser
 import os
@@ -759,7 +758,7 @@ class CertificateAuthorityService(CRUDService):
         ])
 
         if not ca_signed_certs:
-            return ctypes.c_ulong((await self._get_instance(ca_id))['serial']) + 1
+            return int((await self._get_instance(ca_id))['serial']) + 1
         else:
             return max(ca_signed_certs) + 1
